requirements.txt
flickrapi
python-slugify
wget
tqdm
argparse